Rename actor module to message_interface
