+++ Cargo.toml
@@ -47,6 +47,6 @@ ledger-nano = ["iota-client/ledger"]
 ledger-nano-simulator = ["iota-client/ledger"]
 # stronghold = []
 storage = ["rocksdb"]
-actor = []
+message_interface = []
 events = []
 mnemonic = []
