+++ bindings/swift/Cargo.toml
@@ -10,7 +10,7 @@ crate-type = ["cdylib", "staticlib"]      # Creates dynamic lib
 
 [dependencies]
 futures =  { version = "0.3.17", default-features = false }
-iota-wallet = { path = "../../", version = "0.2.0", features = ["actor"] }
+iota-wallet = { path = "../../", version = "0.2.0", features = ["message_interface"] }
 once_cell = { version = "1.8.0", default-features = false }
 serde = { version = "1.0.130", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0.68", default-features = false }
